\name{fPortfolioAdvanced-package}


\alias{fPortfolioAdvanced}


\docType{package}


\title{Advanced Portfolio Optimization}


\description{

    The Rmetrics "fPortfolioAdvanced" is a package which adds 
    advanced functionality to to the Rmetrics "fPortfolio" 
    package

}


\details{

    \tabular{ll}{
        Package:    \tab fPortfolioAdvanced\cr
        Type:       \tab Package\cr
        Date:       \tab 2009\cr
        License:    \tab GPL Version 2 or later\cr
        Copyright:  \tab (c) 1999-2009 Diethelm Wuertz and Rmetrics Association\cr
        URL:        \tab \url{http://www.rmetrics.org}
    }
    
}
    

\keyword{models}

